Criei o comeco dos daos de Produto

+++ petchopp/src/main/java/com/senac/petchopp/interfaces/IDAO.java
@@ -11,7 +11,7 @@ public interface IDAO {
     public void salvar(Object bean);
     public void atualizar(Object bean);
     public void deletar(int id);
-    public Object getById(int id);
+    public Object getById(long id);
     public List<Object> getAll();
     
 }
